change plottype dir name to type and refactor all
