+++ plot_manager/type/ClusterMap.yml
