+++ plot_manager/type/Image.yml
