+++ plot_manager/type/Scatter.yml
