+++ plot_manager/type/cluster_map.py
@@ -1,4 +1,4 @@
-from plot_manager.plottype import plot
+from plot_manager.type import plot
 import seaborn as sb
 import matplotlib.pyplot as plt
 import pandas as pd
