+++ plot_manager/type/pie.py
@@ -1,4 +1,4 @@
-from plot_manager.plottype import anim_plot
+from plot_manager.type import anim_plot
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
